Add function & doc to allow stored-token to be set at drop_auth level

+++ R/drop_auth.R
@@ -21,35 +21,56 @@
 #'@param cache By default your credentials are locally cached in a file called
 #'  \code{.httr-oauth}. Set to FALSE if you need to authenticate separately each
 #'  time.
+#'@param rdstoken File path to stored RDS token. In server environments where
+#'  interactive OAuth is not possible, a token can be created on a desktop
+#'  client and used in production. See \href{https://github.com/karthik/rdrop2}
+#'  for how to create RDS token.
 #'@export
 #'@import httr
 #' @examples \dontrun{
 #' drop_auth()
-#' # If you want to overwrite an existing tokend and switch to a new user,
+#'
+#' # If you want to overwrite an existing local token and switch to a new user,
 #' # set new_user to TRUE.
 #' drop_auth(new_user = TRUE)
+#'
+#' # To use a stored token, set new_user to FALSE and provide token location
+#' drop_auth(new_user = FALSE, rdstoken = "/path/to/tokenfile.RDS")
 #'}
 drop_auth <- function(new_user = FALSE,
                       key = "mmhfsybffdom42w",
                       secret = "l8zeqqqgm1ne5z0",
-                      cache = TRUE) {
+                      cache = TRUE,
+                      rdstoken = NA) {
 
-   if(new_user & file.exists(".httr-oauth")) {
-    message("Removing old credentials ...")
-    file.remove(".httr-oauth")
+  # Check if token file exists & use it
+  if (new_user == FALSE &  !is.na(rdstoken)) {
+    if (file.exists(rdstoken)) {
+      .dstate$token <- readRDS(rdstoken)
+    } else {
+      warning("Token file not found, creating new one ...")
+      new_user <- TRUE
+    }
+  } else {
+    # Force new
+    if (new_user & file.exists(".httr-oauth")) {
+      message("Removing old credentials ...")
+      file.remove(".httr-oauth")
+    }
+
+    # These are the app keys
+    dropbox_app <- httr::oauth_app("dropbox", key, secret)
+
+    # the dropbox endpoint
+    dropbox <- httr::oauth_endpoint(
+      authorize = "https://www.dropbox.com/1/oauth2/authorize",
+      access = "https://api.dropbox.com/1/oauth2/token"
+    )
+    dropbox_token <- httr::oauth2.0_token(dropbox, dropbox_app, cache = cache)
+    stopifnot(inherits(dropbox_token, "Token2.0"))
+
+    .dstate$token <- dropbox_token
   }
-# These are the app keys
-dropbox_app <- httr::oauth_app("dropbox", key, secret)
-
-# the dropbox endpoint
-dropbox <- httr::oauth_endpoint(
-  authorize = "https://www.dropbox.com/1/oauth2/authorize",
-  access = "https://api.dropbox.com/1/oauth2/token"
-)
-dropbox_token <- httr::oauth2.0_token(dropbox, dropbox_app, cache = cache)
-stopifnot(inherits(dropbox_token, "Token2.0"))
-
-.dstate$token <- dropbox_token
 }
 
 
@@ -64,7 +85,7 @@ get_dropbox_token <- function() {
   if(!exists('.dstate') || is.null(.dstate$token)) {
     drop_auth()
   } else {
-.dstate$token
+    .dstate$token
   }
 }
